requirements.txt
langchain
langchain_openai
python-dotenv
streamlit
openai
tiktoken
reportlab